% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootFun.R
\name{bootFun}
\alias{bootFun}
\title{Bootstrapping t-statistics}
\usage{
bootFun(...)
}
\arguments{
\item{...}{Any argument to the function (ignored).}
}
\description{
This function is defunct.
}
\references{
Esarey, Justin, and Jane Lawrence Sumner. 2018. "Corrigendum to 'Marginal Effects in Interaction Models: Determining and Controlling the False Positive Rate.'"
}
