% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMultiLims.R
\name{findMultiLims}
\alias{findMultiLims}
\title{Determine Critical t-Statistic For Marginal Effects Plot}
\usage{
findMultiLims(...)
}
\arguments{
\item{...}{Any argument to the function (ignored).}
}
\description{
This function is defunct.
}
\references{
Esarey, Justin, and Jane Lawrence Sumner. 2018. "Corrigendum to 'Marginal Effects in Interaction Models: Determining and Controlling the False Positive Rate.'"
}
